+++ ui/src/app/auth/auth.module.ts
@@ -39,7 +39,8 @@ import {
 import { AuthRoutingModule } from 'app/auth/auth.routing.module';
 import { RegisterComponent } from 'app/pages/register/register.component';
 import { LogoutComponent } from 'app/pages/logout/logout.component';
-import { RequestPasswordComponent } from 'app/pages/login/request-password.component';
+import { RequestPasswordComponent } from 'app/pages/request-password/request-password.component';
+import { ResetPasswordComponent } from 'app/pages/reset-password/reset-password.component';
 
 @NgModule({
   imports: [
@@ -58,6 +59,7 @@ import { RequestPasswordComponent } from 'app/pages/login/request-password.compo
     AuthComponent,
     LoginComponent,
     RequestPasswordComponent,
+    ResetPasswordComponent,
     RegisterComponent,
     LogoutComponent,
   ],
@@ -66,6 +68,7 @@ import { RequestPasswordComponent } from 'app/pages/login/request-password.compo
     AuthComponent,
     LoginComponent,
     RequestPasswordComponent,
+    ResetPasswordComponent,
     RegisterComponent,
     LogoutComponent,
   ],
