+++ ui/src/app/auth/auth.routing.module.ts
@@ -3,9 +3,9 @@ import { RouterModule, Routes } from '@angular/router';
 import { LoginComponent } from 'app/pages/login/login.component';
 import { LogoutComponent } from 'app/pages/logout/logout.component';
 import { RegisterComponent } from 'app/pages/register/register.component';
-import { NbResetPasswordComponent } from '@nebular/auth';
 import { NgModule } from '@angular/core';
-import { RequestPasswordComponent } from 'app/pages/login/request-password.component';
+import { RequestPasswordComponent } from 'app/pages/request-password/request-password.component';
+import { ResetPasswordComponent } from 'app/pages/reset-password/reset-password.component';
 
 export const routes: Routes = [
   {
@@ -30,7 +30,7 @@ export const routes: Routes = [
   },
   {
     path: 'reset-password',
-    component: NbResetPasswordComponent,
+    component: ResetPasswordComponent,
   },
 ];
