+++ ui/src/app/pages/reset-password/reset-password.component.scss
@@ -1,0 +1,10 @@
+/**
+ * @license
+ * Copyright Akveo. All Rights Reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+:host {
+  .form-group:last-of-type {
+    margin-bottom: 3rem;
+  }
+}
