+++ ui/src/app/pages/reset-password/reset-password.component.ts
@@ -1,0 +1,31 @@
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Inject } from '@angular/core';
+import { Router } from '@angular/router';
+import { NbResetPasswordComponent, NB_AUTH_OPTIONS } from '@nebular/auth';
+
+import { NbAuthService } from '@nebular/auth';
+
+@Component({
+  selector: 'ngx-reset-password-page',
+  styleUrls: ['./reset-password.component.scss'],
+  templateUrl: './reset-password.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class ResetPasswordComponent extends NbResetPasswordComponent {
+
+  redirectDelay: number = 0;
+  showMessages: any = {};
+  strategy: string = '';
+
+  submitted = false;
+  errors: string[] = [];
+  messages: string[] = [];
+  user: any = {};
+
+  constructor(protected service: NbAuthService,
+              @Inject(NB_AUTH_OPTIONS) protected options = {},
+              protected cd: ChangeDetectorRef,
+              protected router: Router) {
+
+    super(service, options, cd, router);
+  }
+}
